Wait for docker daemon to be ready
The build is so short it often gets to the build step before the service is ready. Add a wait so it doesn't try to build before the docker server is listening.

Some of the docker config also updates in line with [latest ACP guidance for docker in Drone](https://ukhomeoffice.github.io/application-container-platform/how-to-docs/index.html#docker-in-docker).